+++ test/bigtest/network/models/campus.js
@@ -1,4 +1,4 @@
-import { Model, hasMany, belongsTo } from '@bigtest/mirage';
+import { Model, hasMany, belongsTo } from 'miragejs';
 
 export default Model.extend({
   libraries: hasMany(),