+++ test/bigtest/network/models/institution.js
@@ -1,6 +1,5 @@
-import { Model, hasMany } from '@bigtest/mirage';
+import { Model, hasMany } from 'miragejs';
 
 export default Model.extend({
-  // mirage is not able to pluralize campus correctly to campuses :(
   campus: hasMany(),
 });